requirements.txt
absl-py==1.2.0
arrow==1.2.2
astunparse==1.6.3
attrs==22.1.0
aws-lambda-builders==1.18.0
aws-sam-cli==1.53.0
aws-sam-translator==1.46.0
binaryornot==0.4.4
boto3==1.24.46
botocore==1.27.46
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
chevron==0.14.0
click==8.1.3
cookiecutter==2.1.1
cycler==0.11.0
dateparser==1.1.1
docker==4.2.2
Flask==2.2.1
Flask-WTF==1.0.1
flatbuffers==1.12
fonttools==4.34.4
gast==0.4.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio @ file:///Users/runner/miniforge3/conda-bld/grpc-split_1658277749120/work
h5py @ file:///Users/runner/miniforge3/conda-bld/h5py_1637964553571/work
idna==3.3
itsdangerous==2.1.2
jedi==0.17.2
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==0.10.0
joblib==1.1.0
jsonschema==3.2.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
MacFSEvents==0.8.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1653325538915/work
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
parso==0.7.1
Pillow==9.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.1
PyYAML==5.4.1
regex==2021.9.30
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.1.1
scipy==1.9.0
serverlessrepo==0.1.10
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-macos==2.9.2
tensorflow-metal==0.5.0
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==3.1.0
tomlkit==0.7.2
typing-extensions==3.10.0.0
tzlocal==3.0
urllib3==1.26.11
watchdog==2.1.2
websocket-client==1.3.3
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
